move types to separate file and consolidate
